test(tests): add missing tests
